VEOSS-2003 - upgrade to scala 3

+++ app/repositories/CachedVatReturnRepository.scala
@@ -24,6 +24,7 @@ import org.mongodb.scala.model._
 import uk.gov.hmrc.mongo.MongoComponent
 import uk.gov.hmrc.mongo.play.json.Codecs.JsonOps
 import uk.gov.hmrc.mongo.play.json.PlayMongoRepository
+import org.mongodb.scala.ObservableFuture
 
 import java.time.{Clock, Instant}
 import java.util.concurrent.TimeUnit
@@ -45,7 +46,7 @@ class CachedVatReturnRepository @Inject()(
         Indexes.ascending("lastUpdated"),
         IndexOptions()
           .name("lastUpdatedIdx")
-          .expireAfter(appConfig.cachedVatReturnTtl, TimeUnit.SECONDS)
+          .expireAfter(appConfig.cachedVatReturnTtl.toLong, TimeUnit.SECONDS)
       ),
       IndexModel(
         Indexes.ascending("userId", "period"),
